Reorganize tests and fix issue with unused warnings
