+++ src/middlewares/mod.rs
@@ -1,1 +1,1 @@
-mod auth;
+pub mod auth;
