+++ src/models/db/mod.rs
@@ -1,1 +1,1 @@
-mod user;
+pub mod user;
