+++ src/models/mod.rs
@@ -1,5 +1,5 @@
 pub mod country;
-mod db;
+pub mod db;
 pub mod event;
 pub mod forms;
 pub mod layers;
